Move typeToTypeKind
